Refactor 7: Status and diagnostics endpoints
